# contracts/openapi/careos-foundation.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ROOTOPATHY CareOS Foundation API",
    "version": "0.21.0",
    "description": "Checked contract for the currently implemented public foundation, browser identity with approved role-level MFA enforcement, approved-registry governed invitations, maker-checker MFA and membership administration, membership-backed organization selection and access projection, and bounded organization-core administration endpoints with the exact approved readiness, organization-profile, registration-identifier, effective address/contact, and versioned international-settings contracts. Unsafe browser operations require an approved Origin or Referer plus the double-submit CSRF cookie/header pair."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1/organizations": {
      "get": {
        "operationId": "listSelectableOrganizations",
        "description": "Returns only organizations for which the authenticated actor has a currently active and effective membership. A stored selection is a UX preference and is not authorization evidence.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Selectable organizations for the current actor",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationAccess"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/v1/organizations/{organizationId}/identifiers": {
      "get": {
        "operationId": "listOrganizationIdentifiers",
        "description": "Returns the minimum-necessary organization identifier projection and live server-authorized actions.",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current governed organization identifier projection",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-CareOS-Session-Expires-In": {
                "$ref": "#/components/headers/SessionExpiresIn"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationIdentifierCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "operationId": "createOrganizationIdentifier",
        "description": "Creates a governed identifier draft with tenant authorization, caller-owned idempotency, an explicit reason, and atomic evidence.",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/BrowserOrigin"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationIdentifierWriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Identifier draft created or exact response replayed",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-CareOS-Session-Expires-In": {
                "$ref": "#/components/headers/SessionExpiresIn"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              },
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationIdentifier"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/identifiers/{identifierId}": {
      "put": {
        "operationId": "updateOrganizationIdentifier",
        "description": "Updates only an identifier draft using its strong revision, caller-owned idempotency, explicit reason, and atomic evidence.",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/IdentifierId"
          },
          {
            "$ref": "#/components/parameters/BrowserOrigin"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationIdentifierWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Identifier draft updated or exact response replayed",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-CareOS-Session-Expires-In": {
                "$ref": "#/components/headers/SessionExpiresIn"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationIdentifier"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/identifiers/{identifierId}/verifications": {
      "post": {
        "operationId": "verifyOrganizationIdentifier",
        "description": "Verifies the exact draft revision with bounded evidence reference, MFA, recent authentication, caller-owned idempotency, and atomic evidence.",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/IdentifierId"
          },
          {
            "$ref": "#/components/parameters/BrowserOrigin"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationIdentifierVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Identifier verified or exact response replayed",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-CareOS-Session-Expires-In": {
                "$ref": "#/components/headers/SessionExpiresIn"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationIdentifier"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/identifiers/{identifierId}/revocations": {
      "post": {
        "operationId": "revokeOrganizationIdentifier",
        "description": "Revokes an eligible verified identifier without rewriting history and preserves any required current primary replacement.",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/IdentifierId"
          },
          {
            "$ref": "#/components/parameters/BrowserOrigin"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationIdentifierRevocationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Identifier revoked or exact response replayed",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-CareOS-Session-Expires-In": {
                "$ref": "#/components/headers/SessionExpiresIn"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationIdentifier"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/identifiers/{identifierId}/supersessions": {
      "post": {
        "operationId": "supersedeOrganizationIdentifier",
        "description": "Atomically supersedes an eligible verified identifier with a separately verified same-type replacement using strong revisions for both records, caller-owned idempotency, and immutable evidence.",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/IdentifierId"
          },
          {
            "$ref": "#/components/parameters/BrowserOrigin"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationIdentifierSupersessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Identifier superseded or exact response replayed",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-CareOS-Session-Expires-In": {
                "$ref": "#/components/headers/SessionExpiresIn"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationIdentifier"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "CAREOS_SESSION",
        "description": "Opaque HttpOnly SameSite=Strict server-side session identifier. Secure is mandatory outside explicit local HTTP development."
      },
      "csrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "XSRF-TOKEN",
        "description": "Readable SameSite=Strict double-submit token cookie. Secure is mandatory outside explicit local HTTP development."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN",
        "description": "Must match the raw XSRF-TOKEN cookie value."
      }
    },
    "parameters": {
      "BrowserOrigin": {
        "name": "Origin",
        "in": "header",
        "required": true,
        "description": "Must exactly match a configured CareOS browser origin. A same-origin Referer is accepted when Origin is unavailable.",
        "schema": {
          "type": "string",
          "format": "uri"
        }
      },
      "OrganizationId": {
        "name": "organizationId",
        "in": "path",
        "required": true,
        "description": "Organization boundary for every protected business-resource route.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Caller-generated key for an explicitly retryable protected mutation. Reuse is valid only for an equivalent request.",
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9._:-]+$"
        }
      }
    },
    "headers": {
      "CorrelationId": {
        "description": "Stable identifier used to correlate one HTTP request across logs and problem responses.",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9._:-]+$"
        }
      },
...
      "ETag": {
        "description": "Strong entity tag for optimistic concurrency.",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^\"[A-Za-z0-9._:-]{1,128}\"$"
        }
      },
      "SessionExpiresIn": {
        "description": "Effective seconds remaining before the authenticated browser session reaches its idle or absolute deadline. Omitted when no authenticated session exists.",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "maximum": 2147483647
        }
      }
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Request validation, password policy, or one-time-token failure",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Credentials, verification evidence, or session is invalid",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Origin, CSRF, or authorization check failed",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource is unavailable or hidden from the current actor",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "The requested transition, idempotency key, or current resource state conflicts with the operation",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "The request failed without exposing sensitive implementation details.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "OrganizationIdentifierCollection": {
        "type": "object",
        "additionalProperties": false,
        "required": ["organizationId", "canCreate", "types", "items"],
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "canCreate": {
            "type": "boolean",
            "description": "True only when the actor has live manage permission and an applicable active type exists."
          },
          "types": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/OrganizationIdentifierType"
            }
          },
          "items": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/OrganizationIdentifier"
            }
          }
        }
      },
      "OrganizationIdentifierWriteRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifierType",
          "assigningAuthority",
          "value",
          "jurisdictionCountryCode",
          "isPrimary",
          "issueDate",
          "expiryDate",
          "effectiveFrom",
          "effectiveTo",
          "reason"
        ],
        "properties": {
          "identifierType": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z][a-z0-9]*([._:-][a-z0-9]+)*$"
          },
          "assigningAuthority": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "value": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "jurisdictionCountryCode": {
            "type": ["string", "null"],
            "pattern": "^[A-Za-z]{2}$"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "issueDate": {
            "type": ["string", "null"],
            "format": "date"
          },
          "expiryDate": {
            "type": ["string", "null"],
            "format": "date"
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date-time"
          },
          "effectiveTo": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "reason": {
            "type": "string",
            "minLength": 10,
            "maxLength": 500
          }
        }
      },
...
      "OrganizationAccess": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "displayName", "status", "roleKeys", "selected"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "status": {
            "type": "string",
            "enum": ["draft", "active"]
          },
          "roleKeys": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          "selected": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}